requirements/common.txt
amqp==2.1.4
anyjson==0.3.3
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
celery==4.0.2
certifi==2018.11.29
chardet==3.0.4
colorclass==2.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.0.11
Django==1.11.20
django-appconf==1.0.2
django-celery-beat==1.0.1
django-celery-results==1.0.1
# remove django-enumfield after squashing migrations
django-enumfield==1.3b2
django-rest-swagger==2.1.2
djangorestframework==3.9.1
dnspython==1.16.0
docopt==0.6.2
drf-nested-routers==0.90.0
flower==0.9.1
gnureadline==6.3.3
inflection==0.3.1
iptools==0.6.1
Jinja2==2.9.6
kombu==4.0.2
MarkupSafe==1.0
netifaces==0.10.6
plop==0.0.28
psycopg2==2.7.1
Pygments==2.2.0
PyYAML==3.12
pytz==2017.2
python-gnupg==0.4.0
redis==2.10.5
requests==2.17.3
tornado==4.5.1
django-compressor==2.1.1
django-crispy-forms==1.6.1
vine==1.1.3
